backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "routing-brain"
version = "0.1.0"
description = "Intelligent LLM Routing Platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "httpx>=0.27.0",
    "anthropic>=0.40.0",
    "openai>=1.35.0",
    "google-genai>=1.0.0",
    "tiktoken>=0.7.0",
    "pyyaml>=6.0.1",
    "sqlalchemy[asyncio]>=2.0.30",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "redis[hiredis]>=5.0.4",
    "prometheus-client>=0.20.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.9",
    "structlog>=24.2.0",
    "tenacity>=8.4.0",
    "orjson>=3.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-httpx>=0.30.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
